+++ app/models/authentication.rb
@@ -1,4 +1,0 @@
-class Authentication < ActiveRecord::Base
-  # Concerns
-  include ContourAuthenticatable
-end