+++ lib/contour/fixes.rb
@@ -1,6 +1,4 @@
 require 'contour/fixes/devise.rb'
-require 'contour/fixes/devise-with-rails-4-2.rb'
-require 'contour/fixes/omniauth.rb'
 require 'contour/fixes/rack.rb'
 require 'contour/fixes/sass-rails.rb'