+++ test/controllers/sessions_controller_test.rb
@@ -21,18 +21,6 @@ class Contour::SessionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should do a graceful redirect to ldap with primary email" do
-    post :create, user: { email: users(:valid).email, password: '' }
-
-    assert_redirected_to '/auth/ldap'
-  end
-
-  test "should do a graceful redirect to google_apps through secondary email" do
-    post :create, user: { email: 'test@gmail.com', password: '' }
-
-    assert_redirected_to '/auth/google_apps'
-  end
-
   test "should not login invalid credentials" do
     post :create, user: { email: '', password: '' }