+++ test/dummy/app/models/user.rb
@@ -7,9 +7,6 @@ class User < ActiveRecord::Base
   devise :database_authenticatable, :registerable, :timeoutable,
          :recoverable, :rememberable, :trackable, :validatable
 
-  # Concerns
-  include Contourable
-
   # # Setup accessible (or protected) attributes for your model
   # attr_accessible :email, :password, :password_confirmation, :remember_me, :first_name, :last_name