+++ test/dummy/db/migrate/20111203203123_devise_create_users.rb
@@ -42,7 +42,7 @@ class DeviseCreateUsers < ActiveRecord::Migration
       # t.string :authentication_token
 
 
-      t.timestamps
+      t.timestamps null: false
     end
 
     add_index :users, :email,                     unique: true