+++ test/models/user_test.rb
@@ -6,8 +6,4 @@ class UserTest < ActiveSupport::TestCase
   test "should get reverse name" do
     assert_equal 'LastName, FirstName', users(:valid).reverse_name
   end
-
-  test "should apply omniauth" do
-    assert_not_nil users(:valid).apply_omniauth({'info' => {'email' => 'Email', 'first_name' => 'FirstName', 'last_name' => 'LastName'}})
-  end
 end